+++ src/Fynance/Yahoo/YDividendResponse.cs
@@ -2,12 +2,12 @@
 
 namespace Fynance.Yahoo
 {
-    internal class YDividendResponse
-    {
-        [JsonProperty("amount")]
-        public double Amount { get; set; }
+	internal class YDividendResponse
+	{
+		[JsonProperty("amount")]
+		public decimal Amount { get; set; }
 
-        [JsonProperty("date")]
-        public double Date { get; set; }
-    }
+		[JsonProperty("date")]
+		public double Date { get; set; }
+	}
 }
