+++ src/Fynance/Yahoo/YIndicatorAdjCloseResponse.cs
@@ -3,9 +3,9 @@ using System.Collections.Generic;
 
 namespace Fynance.Yahoo
 {
-    internal class YIndicatorAdjCloseResponse
-    {
-        [JsonProperty("adjclose")]
-        public IList<double?> AdjClose { get; set; }
-    }
+	internal class YIndicatorAdjCloseResponse
+	{
+		[JsonProperty("adjclose")]
+		public IList<decimal?> AdjClose { get; set; }
+	}
 }
