+++ src/Fynance/Yahoo/YQuoteResponse.cs
@@ -3,21 +3,21 @@ using System.Collections.Generic;
 
 namespace Fynance.Yahoo
 {
-    internal class YQuoteResponse
-    {
-        [JsonProperty("open")]
-        public IList<double?> Open { get; set; }
+	internal class YQuoteResponse
+	{
+		[JsonProperty("open")]
+		public IList<decimal?> Open { get; set; }
 
-        [JsonProperty("high")]
-        public IList<double?> High { get; set; }
+		[JsonProperty("high")]
+		public IList<decimal?> High { get; set; }
 
-        [JsonProperty("low")]
-        public IList<double?> Low { get; set; }
+		[JsonProperty("low")]
+		public IList<decimal?> Low { get; set; }
 
-        [JsonProperty("close")]
-        public IList<double?> Close { get; set; }
+		[JsonProperty("close")]
+		public IList<decimal?> Close { get; set; }
 
-        [JsonProperty("volume")]
-        public IList<double?> Volume { get; set; }
-    }
+		[JsonProperty("volume")]
+		public IList<decimal?> Volume { get; set; }
+	}
 }
